Changed the name from ControlPage and added scale functionality.
